# openapi.yaml
openapi: 3.1.0
info:
  title: Digital Store API
  version: 1.0.0
servers:
  - url: /
security: []
tags:
  - name: catalog
    description: Товары витрины
  - name: orders
    description: Заказы и оплата
  - name: webhooks
    description: События платёжной системы
  - name: providers
    description: Заглушки поставщиков
  - name: admin
    description: Восстановление заказов вручную
  - name: system
    description: Служебные эндпоинты

paths:
  /api/health:
    get:
      tags: [system]
      operationId: health
      summary: Проверить доступность сервиса
      responses:
        '200':
          description: Сервис и база доступны
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'
        '503':
          description: База недоступна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'

  /api/products:
    get:
      tags: [catalog]
      operationId: listProducts
      summary: Получить каталог товаров
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'

  /api/orders:
    post:
      tags: [orders]
      operationId: createOrder
      summary: Создать заказ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Заказ уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Товар раскуплен либо промокод исчерпан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutConflict'

  /api/orders/{orderId}:
    get:
      tags: [orders]
      operationId: getOrder
      summary: Получить заказ
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Заказ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/orders/{orderId}/pay:
    post:
      tags: [orders]
      operationId: payOrder
      summary: Оплатить заказ
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayRequest'
      responses:
        '202':
          description: Вебхук отправлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Заказ нельзя оплатить в текущем статусе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /webhook/payment:
    post:
      tags: [webhooks]
      operationId: receivePaymentWebhook
      summary: Принять уведомление об оплате
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentWebhook'
      responses:
        '200':
          description: Событие принято
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookAccepted'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          description: Событие не принято
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /internal/providers/{provider}/issue:
    post:
      tags: [providers]
      operationId: issueFromProvider
      summary: Запросить код у поставщика
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [a, b]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueRequest'
      responses:
        '200':
          description: Код выдан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '409':
          description: Товара нет в наличии
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueError'
        '503':
          description: Поставщик недоступен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueError'

  /api/admin/orders:
    get:
      tags: [admin]
      operationId: listStuckOrders
      summary: Получить оплаченные, но не выданные заказы
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
            enum: [stuck, all]
            default: stuck
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                type: object
                required: [orders]
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/admin/promocodes:
    get:
      tags: [admin]
      operationId: listPromocodes
      summary: Промокоды с текущим расходом лимита
      responses:
        '200':
          description: Список промокодов
          content:
            application/json:
              schema:
                type: object
                required: [promocodes]
                properties:
                  promocodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Promocode'

  /api/admin/keys:
    post:
      tags: [admin]
      operationId: restockKeys
      summary: Пополнить пул ключей
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestockRequest'
      responses:
        '201':
          description: Ключи добавлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestockResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/admin/orders/{orderId}/retry:
    post:
      tags: [admin]
      operationId: retryDelivery
      summary: Повторить выдачу
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Заказ после попытки выдачи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Заказ нельзя выдать в текущем статусе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  parameters:
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64

  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Не найдено
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Product:
      type: object
      required: [sku, name, type, price, currency]
      properties:
        sku:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [topup, key, subscription, giftcard]
        price:
          type: integer
          minimum: 1
        oldPrice:
          type: [integer, 'null']
          description: Цена до скидки
        currency:
          type: string
        image:
          type: [string, 'null']
        available:
          type: integer
          minimum: 0
          description: Свободных единиц на складе

    OrderStatus:
      type: string
      enum:
        - created
        - paid
        - delivering
        - delivered
        - payment_failed
        - out_of_stock
        - delivery_failed

    Order:
      type: object
      required: [id, sku, amount, discount, total, currency, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        sku:
          type: string
        amount:
          type: integer
          minimum: 0
          description: Цена товара до скидки
        discount:
          type: integer
          minimum: 0
          description: Размер скидки
        total:
          type: integer
          minimum: 0
          description: Сумма к оплате
        currency:
          type: string
        failureReason:
          type: [string, 'null']
        promoCode:
          type: [string, 'null']
        status:
          $ref: '#/components/schemas/OrderStatus'
        code:
          type: [string, 'null']
          description: Код активации товара
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateOrderRequest:
      type: object
      required: [sku, idempotencyKey]
      additionalProperties: false
      properties:
        sku:
          type: string
          minLength: 1
        promoCode:
          type: string
          minLength: 1
          maxLength: 64
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128

    PayRequest:
      type: object
      required: [outcome]
      additionalProperties: false
      properties:
        outcome:
          type: string
          enum: [success, fail]

    PaymentWebhook:
      type: object
      required: [event_id, order_id, status, amount, currency, created_at]
      properties:
        event_id:
          type: string
          minLength: 1
        order_id:
          type: string
          minLength: 1
        status:
          type: string
          enum: [paid, failed]
        amount:
          type: integer
          minimum: 0
          maximum: 2147483647
        currency:
          type: string
        created_at:
          type: string
          format: date-time

    IssueRequest:
      type: object
      required: [request_id, sku, order_id]
      additionalProperties: false
      properties:
        request_id:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        order_id:
          type: string
          minLength: 1

    IssueResponse:
      type: object
      required: [status, request_id, code]
      properties:
        status:
          type: string
          enum: [ok]
        request_id:
          type: string
        code:
          type: string

    IssueError:
      type: object
      required: [status, reason]
      properties:
        status:
          type: string
          enum: [error]
        reason:
          type: string

    CheckoutConflict:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        alternatives:
          type: array
          description: Другие предложения того же типа, доступные к покупке
          items:
            type: object
            required: [sku, name, price, currency, available]
            properties:
              sku:
                type: string
              name:
                type: string
              price:
                type: integer
              currency:
                type: string
              available:
                type: integer

    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum:
            - validation_error
            - not_found
            - conflict
            - promo_limit_reached
            - promo_not_found
            - out_of_stock
            - reservation_expired
            - promo_currency_mismatch
            - product_not_found
            - order_not_payable
            - rate_limited
            - internal_error
        message:
          type: string

    Health:
      type: object
      required: [status, db]
      properties:
        status:
          type: string
          enum: [ok, error]
        db:
          type: string
          enum: [up, down]

    PaymentAccepted:
      type: object
      required: [eventId]
      properties:
        eventId:
          type: string

    WebhookAccepted:
      type: object
      required: [accepted]
      properties:
        accepted:
          type: boolean
        deferred:
          type: boolean

    RestockRequest:
      type: object
      required: [sku, keys]
      additionalProperties: false
      properties:
        sku:
          type: string
          minLength: 1
        keys:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1

    RestockResult:
      type: object
      required: [sku, added, available]
      properties:
        sku:
          type: string
        added:
          type: integer
        available:
          type: integer

    Promocode:
      type: object
      required: [code, type, value, currency, maxUses, usedCount, remaining]
      properties:
        code:
          type: string
        type:
          type: string
          enum: [percent, amount]
        value:
          type: integer
        currency:
          type: [string, 'null']
        maxUses:
          type: integer
        usedCount:
          type: integer
        remaining:
          type: integer

    OrderIdParams:
      type: object
      required: [orderId]
      properties:
        orderId:
          type: string
          minLength: 1
          maxLength: 64

    ProviderParams:
      type: object
      required: [provider]
      properties:
        provider:
          type: string
          enum: [a, b]

    AdminOrdersQuery:
      type: object
      properties:
        state:
          type: string
          enum: [stuck, all]
          default: stuck
          description: stuck — paid, delivering, out_of_stock, delivery_failed